Completed first pass at KSD Opsworks Cookbooks.
This commit represents version 1 of the Chef cookbooks used by
the Kuali Service Delivery team in concert with AWS OpsWorks to
build out the instances in the various OpsWorks stacks.  It
includes:

* the "kinst" base cookbook, containing all of the default
attributes, definitions, and helpers to create services,
and to install components and applications into those services
in a modular and procedural way.  The following components
are currently supported: groovy, jdk, python, ruby, nginx,
jenkins, and tomcat; plus the Kuali Coeus application.

* the "kti" toolkit cookbook, which has the specific attributes
and recipes to support the OpsWorks instance lifecycle for the
KSD stack.

* the "kci" toolkit cookbook, which has the specific attributes
and recipes to support the OpsWorks instance lifecycle for the
UAR (Kuali Coeus) stacks.

The majority of the work is done in "kinst"; the other two are
just very thin wrappers that say, "install these specific
services, components, and application into this stack".

There's a lot left to do.  The "kti" pieces are still very raw,
and still need a lot of polishing before they can take advantage
of the real OpsWOrks cleverness.  The "kci" cookbook is more
complete, but only up to the point of installing the KC app --
most of that definition is still just stubbed out comments.

This commit also marks the point at which Mike is letting go
of the code, and transitioning further development efforts to
the Kuali Service Delivery team.